#8 Separate system tests from unit tests and run in separate Jenkins node
